Remove paypal buttons from venmoPayments
