--- client/components/venmoPayments/oneTimePayment/html/src/custom/sandboxedIframe/.npmrc
@@ -1,2 +1,0 @@
-registry=https://registry.npmjs.org
-package-lock=false
