--- client/components/venmoPayments/oneTimePayment/html/src/custom/sandboxedIframe/.nvmrc
@@ -1,1 +1,0 @@
-20.18.1
