--- client/components/venmoPayments/oneTimePayment/html/src/custom/sandboxedIframe/src/merchant-example/index.html
@@ -1,159 +1,0 @@
-<!doctype html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <title>PayPal iframe Example</title>
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-
-    <link
-      rel="icon"
-      href="data:image/svg+xml,<svg xmlns=%22http://www.w3.org/2000/svg%22 viewBox=%220 0 100 100%22><text y=%22.9em%22 font-size=%2290%22>💸</text></svg>"
-    />
-
-    <style>
-      :root {
-        --theme-color-text: #0a1b0f;
-        --theme-color-background: #f7fcf9;
-        --theme-color-primary: #334037;
-        --theme-color-secondary: #93e3ab;
-        --theme-color-accent: #62dc87;
-        --theme-color-overlay-cta: #ffffff;
-        --theme-color-overlay-background: #000000aa;
-        --theme-color-overlay-background-close-text: #cccccc;
-      }
-
-      body {
-        margin: 0;
-        font-family: sans-serif;
-        color: var(--theme-color-text);
-      }
-
-      button {
-        border: none;
-        font-size: 1em;
-        padding: 10px;
-        border-radius: 5px;
-      }
-
-      #stateDebug {
-        padding: 5px;
-        position: fixed;
-        bottom: 0;
-        left: 0;
-        background: var(--theme-color-secondary);
-        z-index: 100;
-      }
-
-      #mainContainer {
-        padding: 10px;
-      }
-
-      #extraHeight {
-        background: var(--theme-color-background);
-        height: 1000px;
-        width: 100%;
-      }
-
-      #overlayContainer {
-        background: none;
-        border: none;
-        overflow: visible;
-        text-align: center;
-      }
-
-      #overlayContainerLogo {
-        margin-bottom: 10px;
-      }
-
-      #overlayCloseButtonCTA {
-        background: var(--theme-color-overlay-cta);
-        cursor: pointer;
-        border-radius: 20px;
-        padding: 10px 20px;
-      }
-
-      #overlayCloseButtonBackdrop {
-        background: none;
-        color: var(--theme-color-overlay-background-close-text);
-        position: fixed;
-        top: 0;
-        right: 0;
-        font-size: 1.5em;
-        cursor: pointer;
-      }
-
-      #overlayContainer::backdrop {
-        background: var(--theme-color-overlay-background);
-      }
-
-      #iframeWrapper {
-        border: none;
-        width: 100%;
-        height: 300px;
-      }
-
-      #iframeWrapper.fullWindow {
-        position: absolute;
-        top: 0;
-        left: 0;
-        height: 100%;
-        width: 100%;
-      }
-
-      .stateValue {
-        font-family: monospace;
-      }
-    </style>
-  </head>
-  <body>
-    <script async src="/app.js" onload="onLoad()"></script>
-
-    <div id="mainContainer">
-      <h1>PayPal iframe Example</h1>
-
-      <p>
-        The parent page is an example merchant page. Within the merchant page,
-        an iframe uses the PayPal web SDK and communicates with a server that
-        has access to the PayPal server SDK. The merchant page and iframe then
-        communicate about the state of the payment flow via postMessage. Using
-        the PayPal web SDK in this way, inside an iframe, avoids having to
-        request the SDK script directly from the merchant page.
-      </p>
-
-      <div id="stateDebug">
-        <div>Merchant Page State</div>
-        <li>
-          <span>merchant page domain:</span>
-          <span id="merchantDomain" class="stateValue"></span>
-        </li>
-        <li>
-          <span>last message from iframe: </span>
-          <span id="postMessageStatus" class="stateValue">No messages 😴</span>
-        </li>
-        <li>
-          <span>selected presentationMode from iframe: </span>
-          <span id="presentationMode" class="stateValue">Not set 🤔</span>
-        </li>
-      </div>
-
-      <iframe
-        id="iframeWrapper"
-        src="http://localhost:3000/?origin=http://localhost:3001"
-        sandbox="allow-scripts allow-same-origin allow-popups allow-forms"
-        allow="payment"
-      ></iframe>
-
-      <div id="extraHeight">Some extra height follows...</div>
-    </div>
-
-    <dialog id="overlayContainer">
-      <div id="overlayContainerLogo">
-        <img
-          src="https://www.paypalobjects.com/js-sdk-logos/2.2.7/paypal-white.svg"
-        />
-      </div>
-      <button id="overlayCloseButtonCTA">Close</button>
-      <button id="overlayCloseButtonBackdrop">X</button>
-    </dialog>
-  </body>
-</html>
