--- client/components/venmoPayments/oneTimePayment/html/src/custom/sandboxedIframe/src/paypal-iframe/index.html
@@ -1,95 +1,0 @@
-<!doctype html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <title>PayPal Payment Handler</title>
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-
-    <style>
-      :root {
-        --theme-color-text: #e5f5ea;
-        --theme-color-background: #334037;
-        --theme-color-primary: #becbc2;
-        --theme-color-secondary: #1c6d36;
-        --theme-color-accent: #239f48;
-      }
-
-      html {
-        background: var(--theme-color-background);
-      }
-
-      body {
-        font-family: sans-serif;
-        color: var(--theme-color-text);
-      }
-
-      .stateValue {
-        font-family: monospace;
-      }
-
-      #config {
-        padding-bottom: 10px;
-      }
-    </style>
-  </head>
-  <body>
-    <div id="stateDebug">
-      <div>Iframe Page State</div>
-      <ul>
-        <li>
-          <span>iframe domain:</span>
-          <span id="iframeDomain" class="stateValue"></span>
-        </li>
-        <li>
-          <span>last message from parent: </span>
-          <span id="postMessageStatus" class="stateValue">No messages 😴</span>
-        </li>
-      </ul>
-    </div>
-
-    <div id="config">
-      <div>Payment flow config</div>
-      <fieldset>
-        <legend>presentationMode</legend>
-        <div>
-          <input
-            checked
-            type="radio"
-            id="presentationMode-popup"
-            name="presentationMode"
-            value="popup"
-          />
-          <label for="presentationMode-popup">popup</label>
-        </div>
-        <div>
-          <input
-            type="radio"
-            id="presentationMode-modal"
-            name="presentationMode"
-            value="modal"
-          />
-          <label for="presentationMode-modal">modal</label>
-        </div>
-        <div>
-          <input
-            type="radio"
-            id="presentationMode-payment-handler"
-            name="presentationMode"
-            value="payment-handler"
-          />
-          <label for="presentationMode-payment-handler">payment-handler</label>
-        </div>
-      </fieldset>
-    </div>
-
-    <paypal-button id="paypal-button"></paypal-button>
-
-    <script src="/integration.js"></script>
-
-    <script
-      async
-      src="https://www.sandbox.paypal.com/web-sdk/v6/core"
-      onload="onPayPalLoaded()"
-    ></script>
-  </body>
-</html>
