--- client/components/venmoPayments/oneTimePayment/html/src/custom/sandboxedIframe/vite-merchant-example.config.js
@@ -1,10 +1,0 @@
-import { defineConfig } from "vite";
-
-// https://vitejs.dev/config/
-export default defineConfig({
-  plugins: [],
-  root: "src/merchant-example",
-  server: {
-    port: 3001,
-  },
-});
