--- client/components/venmoPayments/oneTimePayment/html/src/custom/sandboxedIframe/vite-paypal-iframe.config.js
@@ -1,17 +1,0 @@
-import { defineConfig } from "vite";
-
-// https://vitejs.dev/config/
-export default defineConfig({
-  plugins: [],
-  root: "src/paypal-iframe",
-  server: {
-    port: 3000,
-    proxy: {
-      "/paypal-api": {
-        target: "http://localhost:8080",
-        changeOrigin: true,
-        secure: false,
-      },
-    },
-  },
-});
