+++ client/components/venmoPayments/oneTimePayment/html/src/index.html
@@ -2,24 +2,28 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <title>One-Time Payment Sample Integrations - PayPal Web SDK</title>
+    <title>One-Time Payment - Recommended Integration - PayPal Web SDK</title>
     <meta name="viewport" content="width=device-width, initial-scale=1" />
+    <style>
+      .buttons-container {
+        display: flex;
+        flex-direction: column;
+        gap: 12px;
+      }
+    </style>
   </head>
   <body>
-    <h1>One-Time Payment Integrations</h1>
+    <h1>One-Time Payment Recommended Integration</h1>
 
-    <ul>
-      <li>
-        <a href="/recommended/index.html"
-          >Recommended integration with PayPal, Venmo, PayLater, and PayPal
-          Credit buttons</a
-        >
-      </li>
-      <li>
-        <a href="/custom/paymentHandler/index.html"
-          >Custom "payment-handler" presentation mode with PayPal button</a
-        >
-      </li>
-    </ul>
+    <div class="buttons-container">
+      <venmo-button id="venmo-button" type="pay" hidden></venmo-button>
+    </div>
+    <script src="app.js"></script>
+
+    <script
+      async
+      src="https://www.sandbox.paypal.com/web-sdk/v6/core"
+      onload="onPayPalLoaded()"
+    ></script>
   </body>
 </html>
