--- client/components/venmoPayments/oneTimePayment/html/src/recommended/app.js
@@ -1,181 +1,0 @@
-async function onPayPalLoaded() {
-  try {
-    const clientToken = await getBrowserSafeClientToken();
-    const sdkInstance = await window.paypal.createInstance({
-      clientToken,
-      components: ["paypal-payments", "venmo-payments"],
-      pageType: "checkout",
-    });
-
-    const paymentMethods = await sdkInstance.findEligibleMethods({
-      currencyCode: "USD",
-    });
-
-    if (paymentMethods.isEligible("paypal")) {
-      setupPayPalButton(sdkInstance);
-    }
-
-    if (paymentMethods.isEligible("venmo")) {
-      setupVenmoButton(sdkInstance);
-    }
-
-    if (paymentMethods.isEligible("paylater")) {
-      const paylaterPaymentMethodDetails =
-        paymentMethods.getDetails("paylater");
-      setupPayLaterButton(sdkInstance, paylaterPaymentMethodDetails);
-    }
-
-    if (paymentMethods.isEligible("credit")) {
-      const paypalCreditPaymentMethodDetails =
-        paymentMethods.getDetails("credit");
-      setupPayPalCreditButton(sdkInstance, paypalCreditPaymentMethodDetails);
-    }
-  } catch (error) {
-    console.error(error);
-  }
-}
-
-const paymentSessionOptions = {
-  async onApprove(data) {
-    console.log("onApprove", data);
-    const orderData = await captureOrder({
-      orderId: data.orderId,
-    });
-    console.log("Capture result", orderData);
-  },
-  onCancel(data) {
-    console.log("onCancel", data);
-  },
-  onError(error) {
-    console.log("onError", error);
-  },
-};
-
-async function setupPayPalButton(sdkInstance) {
-  const paypalPaymentSession = sdkInstance.createPayPalOneTimePaymentSession(
-    paymentSessionOptions,
-  );
-
-  const paypalButton = document.querySelector("#paypal-button");
-  paypalButton.removeAttribute("hidden");
-
-  paypalButton.addEventListener("click", async () => {
-    try {
-      await paypalPaymentSession.start(
-        { presentationMode: "auto" },
-        createOrder(),
-      );
-    } catch (error) {
-      console.error(error);
-    }
-  });
-}
-
-async function setupVenmoButton(sdkInstance) {
-  const venmoPaymentSession = sdkInstance.createVenmoOneTimePaymentSession(
-    paymentSessionOptions,
-  );
-  const venmoButton = document.querySelector("#venmo-button");
-  venmoButton.removeAttribute("hidden");
-
-  venmoButton.addEventListener("click", async () => {
-    try {
-      await venmoPaymentSession.start(
-        { presentationMode: "auto" },
-        createOrder(),
-      );
-    } catch (error) {
-      console.error(error);
-    }
-  });
-}
-
-async function setupPayLaterButton(sdkInstance, paylaterPaymentMethodDetails) {
-  const paylaterPaymentSession =
-    sdkInstance.createPayLaterOneTimePaymentSession(paymentSessionOptions);
-
-  const { productCode, countryCode } = paylaterPaymentMethodDetails;
-  const paylaterButton = document.querySelector("#paylater-button");
-
-  paylaterButton.productCode = productCode;
-  paylaterButton.countryCode = countryCode;
-  paylaterButton.removeAttribute("hidden");
-
-  paylaterButton.addEventListener("click", async () => {
-    try {
-      await paylaterPaymentSession.start(
-        { presentationMode: "auto" },
-        createOrder(),
-      );
-    } catch (error) {
-      console.error(error);
-    }
-  });
-}
-
-async function setupPayPalCreditButton(
-  sdkInstance,
-  paypalCreditPaymentMethodDetails,
-) {
-  const paypalCreditPaymentSession =
-    sdkInstance.createPayPalCreditOneTimePaymentSession(paymentSessionOptions);
-
-  const { countryCode } = paypalCreditPaymentMethodDetails;
-  const paypalCreditButton = document.querySelector("#paypal-credit-button");
-
-  paypalCreditButton.countryCode = countryCode;
-  paypalCreditButton.removeAttribute("hidden");
-
-  paypalCreditButton.addEventListener("click", async () => {
-    try {
-      await paypalCreditPaymentSession.start(
-        { presentationMode: "auto" },
-        createOrder(),
-      );
-    } catch (error) {
-      console.error(error);
-    }
-  });
-}
-
-async function getBrowserSafeClientToken() {
-  const response = await fetch("/paypal-api/auth/browser-safe-client-token", {
-    method: "GET",
-    headers: {
-      "Content-Type": "application/json",
-    },
-  });
-  const { accessToken } = await response.json();
-
-  return accessToken;
-}
-
-async function createOrder() {
-  const response = await fetch(
-    "/paypal-api/checkout/orders/create-with-sample-data",
-    {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-    },
-  );
-  const { id } = await response.json();
-
-  return { orderId: id };
-}
-
-async function captureOrder({ orderId }) {
-  const response = await fetch(
-    `/paypal-api/checkout/orders/${orderId}/capture`,
-    {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-    },
-  );
-  const data = await response.json();
-
-  return data;
-}
